project-back-api/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql  # Si usas MySQL como base de datos
asyncpg  # Si usas PostgreSQL
pyserial